API/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
databricks-sql-connector>=3.0.0
